+++ src/main/java/com/example/jobsearch/service/ResumeService.java
@@ -10,10 +10,14 @@ import com.example.jobsearch.model.Resume;
 import com.example.jobsearch.model.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
+
 @Slf4j
 @Service
 @RequiredArgsConstructor
@@ -23,15 +27,15 @@ public class ResumeService {
     private final WorkExperienceService workExperienceService;
     private final ContactInfoService contactInfoService;
     private final ApplicantProfileService applicantService;
-    private final CategoryService categoryService;
     private final UserService userService;
+    private final CategoryService categoryService;
 
     private ResumeDto makeDtoFromResume(Resume resume) {
         return ResumeDto.builder()
                 .id(resume.getId())
-                .applicantDto(applicantService.getApplicantById(resume.getApplicantId()))
+                .authorEmail(applicantService.getApplicantById(resume.getApplicantId()).getUserId())
                 .resumeTitle(resume.getResumeTitle())
-                .categoryDto(categoryService.getCategoryByName(resume.getCategory()))
+                .category(resume.getCategory())
                 .expectedSalary(resume.getExpectedSalary())
                 .isActive(resume.isActive())
                 .isPublished(resume.isPublished())
@@ -41,9 +45,9 @@ public class ResumeService {
     private Resume createResumeFromDto(ResumeDto resume) {
         return Resume.builder()
                 .id(resume.getId())
-                .applicantId(resume.getApplicantDto().getId())
+                .applicantId(applicantService.findApplicantByUserId(resume.getAuthorEmail()).getId())
                 .resumeTitle(resume.getResumeTitle())
-                .category(resume.getCategoryDto().getCategory())
+                .category(resume.getCategory())
                 .expectedSalary(resume.getExpectedSalary())
                 .isActive(resume.isActive())
                 .isPublished(resume.isPublished())
@@ -72,29 +76,73 @@ public class ResumeService {
                 .toList();
     }
 
-    public List<ResumeDto> getAllResumesByCategoryName(String categoryName) {
-        List<Resume> list = resumeDao.getAllResumesByCategoryName(categoryName);
-        return list.stream()
-                .map(this::makeDtoFromResume)
-                .toList();
-    }
-
-    public void createResume(ResumeDto e, Authentication auth) {
+    public ResponseEntity<?> createResume(ResumeDto e, Authentication auth) {
         var u = auth.getPrincipal();
         User user = userService.getUserFromAuth(u.toString());
-        if(user.getId().equalsIgnoreCase(e.getApplicantDto().getUserId())) {
-            log.warn("New resume created: {}", e.getId());
-            resumeDao.save(createResumeFromDto(e));
+        if(user.getId().equalsIgnoreCase(e.getAuthorEmail())) {
+            Optional<Resume> r;
+            if(e.getId() == null) {
+            long x = (long)resumeDao.getAllResumes().size()+1;
+            r = resumeDao.getResumeById(x);
+            } else {
+                r = resumeDao.getResumeById(e.getId());
+            }
+            if(r.isEmpty()) {
+                if(categoryService.getCategory(e.getCategory()).isPresent()) {
+                    resumeDao.save(createResumeFromDto(e));
+                    return new ResponseEntity<>("Resume created successfully", HttpStatus.OK);
+                } else {
+                    log.warn("Tried to use a category that does not exist: {}", e.getCategory());
+                    return new ResponseEntity<>("Category does not exist", HttpStatus.BAD_REQUEST);
+                }
+            } else {
+                log.info("Tried to create a resume that already exists: {}", e.getId());
+                return new ResponseEntity<>("Resume already exists", HttpStatus.OK);
+            }
+        } else {
+            log.warn("Tried to create a resume for another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to create a resume for another user", HttpStatus.BAD_REQUEST);
+        }
+    }
+    public ResponseEntity<?> deleteResume(ResumeDto e, Authentication auth) {
+        var u = auth.getPrincipal();
+        User user = userService.getUserFromAuth(u.toString());
+        if(user.getId().equalsIgnoreCase(e.getAuthorEmail())) {
+            var r = resumeDao.getResumeById(e.getId());
+            if (r.isPresent()) {
+                resumeDao.delete(e.getId());
+                return new ResponseEntity<>("Resume deleted successfully", HttpStatus.OK);
+            } else {
+                log.info("Tried to delete a resume that does not exist: {}", e.getId());
+                return new ResponseEntity<>("Tried to delete a resume that does not exist", HttpStatus.OK);
+            }
+        } else {
+            log.warn("Tried to delete a resume of another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to delete a resume of another user", HttpStatus.BAD_REQUEST);
         }
     }
 
-    public void deleteResume(ResumeDto e) {
-        log.warn("Resume deleted: {}", e.getId());
-        resumeDao.delete(e.getId());
-    }
-
-    public void editResume(ResumeDto e) {
-        resumeDao.editResume(createResumeFromDto(e));
+    public ResponseEntity<?> editResume(ResumeDto e, Authentication auth) {
+        var u = auth.getPrincipal();
+        User user = userService.getUserFromAuth(u.toString());
+        if(user.getId().equalsIgnoreCase(e.getAuthorEmail())) {
+            var r = resumeDao.getResumeById(e.getId());
+            if (r.isPresent()) {
+                if(categoryService.getCategory(e.getCategory()).isPresent()) {
+                    resumeDao.editResume(createResumeFromDto(e));
+                    return new ResponseEntity<>("Resume edited successfully", HttpStatus.OK);
+                }else {
+                    log.warn("Tried to use a category that does not exist: {}", e.getCategory());
+                    return new ResponseEntity<>("Category does not exist", HttpStatus.BAD_REQUEST);
+                }
+        } else {
+                log.warn("Tried to edit a resume that does not exist: {}", e.getId());
+                return new ResponseEntity<>("Cannot edit a resume that does not exist", HttpStatus.NOT_FOUND);
+            }
+        } else {
+            log.warn("Tried to edit a resume of another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to edit a resume of another user", HttpStatus.BAD_REQUEST);
+        }
     }
 
     public List<ResumeDto> getAllResumesByLongList(List<Long> list) {
@@ -105,7 +153,7 @@ public class ResumeService {
     }
 
     public ResumeDto getResumeById(long id) {
-        return makeDtoFromResume(resumeDao.getResumeById(id));
+        return makeDtoFromResume(resumeDao.findResumeById(id));
     }
 
     public List<ResumeDto> getResumeByResumeTitle(String title) {
@@ -115,40 +163,41 @@ public class ResumeService {
                 .toList();
     }
 
-    public void addEducation(EducationDto educationDto) {
-        educationService.createEducation(educationDto);
+    public ResponseEntity<?> addEducation(EducationDto educationDto) {
+        return educationService.createEducation(educationDto);
+
     }
 
-    public void editEducation(EducationDto educationDto) {
-        educationService.editEducation(educationDto);
+    public ResponseEntity<?> editEducation(EducationDto educationDto) {
+        return educationService.editEducation(educationDto);
     }
 
-    public void deleteEducation(EducationDto educationDto) {
-        educationService.deleteEducation(educationDto);
+    public ResponseEntity<?>  deleteEducation(EducationDto educationDto) {
+        return educationService.deleteEducation(educationDto);
     }
 
-    public void addWorkExperience(WorkExperienceDto workExperienceDto) {
-        workExperienceService.createWorkExperience(workExperienceDto);
+    public ResponseEntity<?> addWorkExperience(WorkExperienceDto workExperienceDto) {
+        return workExperienceService.createWorkExperience(workExperienceDto);
     }
 
-    public void editWorkExperience(WorkExperienceDto workExperienceDto) {
-        workExperienceService.editWorkExperience(workExperienceDto);
+    public ResponseEntity<?> editWorkExperience(WorkExperienceDto workExperienceDto) {
+        return workExperienceService.editWorkExperience(workExperienceDto);
     }
 
-    public void deleteWorkExperience(WorkExperienceDto workExperienceDto) {
-        workExperienceService.deleteWorkExperience(workExperienceDto);
+    public ResponseEntity<?> deleteWorkExperience(WorkExperienceDto workExperienceDto) {
+        return workExperienceService.deleteWorkExperience(workExperienceDto);
     }
 
-    public void addContactInfo(ContactInfoDto contactInfoDto) {
-        contactInfoService.createContactInfo(contactInfoDto);
+    public ResponseEntity<?> addContactInfo(ContactInfoDto contactInfoDto) {
+        return contactInfoService.createContactInfo(contactInfoDto);
     }
 
-    public void editContactInfo(ContactInfoDto contactInfoDto) {
-        contactInfoService.editContactInfo(contactInfoDto);
+    public ResponseEntity<?>  editContactInfo(ContactInfoDto contactInfoDto) {
+        return contactInfoService.editContactInfo(contactInfoDto);
     }
 
-    public void deleteContactInfo(ContactInfoDto contactInfoDto) {
-        contactInfoService.deleteContactInfo(contactInfoDto);
+    public ResponseEntity<?>  deleteContactInfo(ContactInfoDto contactInfoDto) {
+        return contactInfoService.deleteContactInfo(contactInfoDto);
     }
     public List<WorkExperienceDto> getAllWorkExperienceByResumeId(long resumeId) {
         return workExperienceService.getAllWorkExperienceByResumeId((resumeId));
